Refactor Create Tenant form into shared TenantFormPage and utils.
Extract types, parse/build/validate helpers, and a mode-aware form
component. CreateTenantPage is now a thin create-mode wrapper with no
behaviour change to /tenant-create.
